+++ actionpack/CHANGELOG.md
@@ -1,4 +1,8 @@
-*   Remove deprecated support passing path as `nil` in `ActionDispatch::Routing::Mapper#match`.
+*   Remove deprecated support for passing `:path` and route path as stings in `ActionDispatch::Routing::Mapper#match`.
+
+    *Rafael Mendonça França*
+
+*   Remove deprecated support for passing path as `nil` in `ActionDispatch::Routing::Mapper#match`.
 
     *Rafael Mendonça França*